Publish the delta algorithm seam and address inputs in 64 bits (#26)
DeltaAlgorithm, DeltaScanner and DeltaSink are public, and the patcher
constructors take one, so a third party can replace how bytes are matched
without touching the patch format. Nothing about the search reaches the
container - the COPY/DIFF/ADD/RUN set is fixed - so a patch built by a
custom algorithm is restored by the stock patcher.

Inputs are now a SeekableSource rather than a ByteArray. A delta is not a
sequential transform: a copy names an arbitrary source offset and applying
one reads it back, so the right abstraction is random access, not a stream.
fileSource(path) opens a file for it on every target with a file system -
RandomAccessFile on the JVM, stdio on native, fs.readSync on Node - and the
browser, which has none, says so. Checksums now run straight off a source a
chunk at a time, so a file is no longer read whole to be verified.

Sizes and delta offsets become 64-bit, format v2. For values inside Int
range the varint bytes are exactly what v1 wrote, so v1 patches still
apply; only the declared ceiling moved.

The bundled RollingHashAlgorithm still indexes a ByteArray, so a patcher
materializes both inputs once before encoding and refuses anything above
2 GB with a clear error. A bounded-memory search is the next step and needs
no format change when it lands.

+++ library/src/commonTest/kotlin/com/linroid/kiff/PatchRoundTripSupport.kt
@@ -9,8 +9,8 @@ fun assertRestores(patcher: Patcher, source: ByteArray, target: ByteArray): Int 
   val patch = patcher.createPatch(source, target)
   val info = Kiff.info(patch)
   assertEquals(patcher.id, info.patcher)
-  assertEquals(source.size, info.sourceSize)
-  assertEquals(target.size, info.targetSize)
+  assertEquals(source.size.toLong(), info.sourceSize)
+  assertEquals(target.size.toLong(), info.targetSize)
   assertContentEquals(target, patcher.applyPatch(source, patch), "restored bytes differ")
   return patch.size
 }
